Fix incorrect handling of polyfills
- The System.Runtime.CompilerServices / System.Diagnostics.CodeAnalysis types should all be internal rather than public; otherwise they will lead to ambiguity warnings with consumers and also won't type unify.
- The System.Collection.Immutable / ValueTask types should not be polyfilled; they instead should just use the netstandard2.0 packages.
- Anthropic.Foundry needs to include the shims.cs file correctly.
